shell/requirements.txt
flask>=3.0
gunicorn>=22.0
psycopg2-binary>=2.9
requests>=2.31
cryptography>=42.0.0
-e ../shared
